+++ com.unity.visualeffectgraph/Editor/Data/VFXDataParticle.cs
@@ -741,8 +741,11 @@ namespace UnityEditor.VFX
                 }
 
                 uniformMappings.Clear();
+                
                 foreach (var uniform in contextData.uniformMapper.uniforms)
                     uniformMappings.Add(new VFXMapping(contextData.uniformMapper.GetName(uniform), expressionGraph.GetFlattenedIndex(uniform)));
+                foreach (var buffer in contextData.uniformMapper.buffers)
+                    uniformMappings.Add(new VFXMapping(contextData.uniformMapper.GetName(buffer), expressionGraph.GetFlattenedIndex(buffer)));
                 foreach (var texture in contextData.uniformMapper.textures)
                 {
                     // TODO At the moment issue all names sharing the same texture as different texture slots. This is not optimized as it required more texture binding than necessary